functions/purpleair_connector/requirements.txt
pandas
matplotlib
geojson
google-cloud-bigquery
google-cloud-firestore